task5-mini-project/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
feedparser>=6.0.10
schedule>=1.2.0
python-dotenv>=1.0.0
pytest>=7.0.0
